backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq>=0.30.0,<1.0.0
yfinance==0.2.50
pandas==2.2.3
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.3
snowflake-connector-python==3.12.3
finnhub-python==2.4.20
